(fix) List must exist to add/delete items

+++ src/commands/item/options/delete/handler.ts
@@ -1,5 +1,6 @@
 import { InteractionResponseType } from 'discord-interactions';
 import { deleteItem } from '../../../../db';
+import { getList } from '../../../../db/getList';
 import { OptionHandler } from '../../../../types/discord';
 import { getOwnerID } from '../../../../utils/watchlist';
 import { ItemCommand } from '../../cmd';
@@ -11,6 +12,15 @@ export const handler: OptionHandler<Option, ItemCommand> = async (opt, cmd) => {
 
   const list = params[0].value.trim();
   const itemNum = params[1].value;
+
+  const listRes = await getList({ ownerID, name: list });
+
+  if (!listRes.success) {
+    return {
+      type: InteractionResponseType.CHANNEL_MESSAGE_WITH_SOURCE,
+      data: { content: listRes.message },
+    };
+  }
 
   const res = await deleteItem({ list, itemNum, ownerID });
   if (!res.success) {
